requirements/prod.txt
Flask==0.10.1
Flask-Assets==0.8
Flask-Login==0.2.9
Flask-Migrate==1.2.0
Flask-OAuthlib==0.4.2
Flask-SQLAlchemy==1.0
Flask-Script==0.6.6
Flask-WTF==0.9.4
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.18
SQLAlchemy==0.9.1
WTForms==1.0.5
Werkzeug==0.9.4
alembic==0.6.2
argparse==1.2.1
cssmin==0.2.0
eveapi==1.2.6
gunicorn==18.0
hiredis==0.1.2
itsdangerous==0.23
jsmin==2.0.8
oauthlib==0.6.1
python-ldap==2.4.13
redis==2.9.1
webassets==0.8
wsgiref==0.1.2
